+++ mayan/apps/document_states/tests/test_workflow_template_transition_field_views.py
@@ -1,3 +1,5 @@
+from django.utils import timezone
+
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
 from ..events import (
@@ -194,6 +196,52 @@ class WorkflowTemplateTransitionFieldViewTestCase(
         self.assertContains(
             response=response, status_code=200,
             text=self._test_workflow_template_transition_field.label
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+
+class WorkflowTemplateTransitionFieldDefaultViewTestCase(
+    WorkflowInstanceViewTestMixin,
+    WorkflowTemplateTransitionFieldViewTestMixin, GenericViewTestCase
+):
+    def setUp(self):
+        super().setUp()
+        self._create_test_workflow_template(add_test_document_type=True)
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition()
+
+    def test_lookup_field(self):
+        self._create_test_workflow_template_transition_field(
+            extra_data={'default': 'TEST-{{ now.year }}-TEST'}
+        )
+
+        self._create_test_document_stub()
+
+        self.grant_access(
+            obj=self._test_document,
+            permission=permission_workflow_instance_transition
+        )
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_instance_transition
+        )
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_transition_execute_get_view()
+        self.assertContains(
+            response=response, text='TEST-{}-TEST'.format(
+                timezone.now().year
+            ),
+            status_code=200
+        )
+
+        self.assertEqual(
+            self._test_workflow_instance.get_current_state(),
+            self._test_workflow_template_state_list[0]
         )
 
         events = self._get_test_events()